:recycle: Move factory tests to factory package
